# feat: use faster query_and_wait method from google-cloud-bigquery when available