--- a/pandas_gbq/features.py
+++ b/pandas_gbq/features.py
@@ -4,8 +4,9 @@
 
 """Module for checking dependency versions and supported features."""
 
-# https://github.com/googleapis/python-bigquery/blob/master/CHANGELOG.md
+# https://github.com/googleapis/python-bigquery/blob/main/CHANGELOG.md
 BIGQUERY_MINIMUM_VERSION = "3.3.5"
+BIGQUERY_QUERY_AND_WAIT_VERSION = "3.14.0"
 PANDAS_VERBOSITY_DEPRECATION_VERSION = "0.23.0"
 PANDAS_BOOLEAN_DTYPE_VERSION = "1.0.0"
 PANDAS_PARQUET_LOSSLESS_TIMESTAMP_VERSION = "1.1.0"
@@ -44,6 +45,13 @@ class Features:
             )
 
         return google.cloud.bigquery
+
+    @property
+    def bigquery_has_query_and_wait(self):
+        import packaging.version
+
+        min_version = packaging.version.parse(BIGQUERY_QUERY_AND_WAIT_VERSION)
+        return self.bigquery_installed_version >= min_version
 
     @property
     def pandas_installed_version(self):