--- a/tests/unit/test_context.py
+++ b/tests/unit/test_context.py
@@ -8,6 +8,7 @@ from unittest import mock
 
 import google.cloud.bigquery
 import google.cloud.bigquery.table
+import packaging.version
 import pytest
 
 
@@ -55,8 +56,15 @@ def test_read_gbq_should_save_credentials(mock_get_credentials):
     mock_get_credentials.assert_not_called()
 
 
-def test_read_gbq_should_use_dialect(mock_bigquery_client):
+def test_read_gbq_should_use_dialect_with_query(monkeypatch, mock_bigquery_client):
     import pandas_gbq
+    import pandas_gbq.features
+
+    monkeypatch.setattr(
+        pandas_gbq.features.FEATURES,
+        "_bigquery_installed_version",
+        packaging.version.parse(pandas_gbq.features.BIGQUERY_MINIMUM_VERSION),
+    )
 
     assert pandas_gbq.context.dialect is None
     pandas_gbq.context.dialect = "legacy"
@@ -69,5 +77,37 @@ def test_read_gbq_should_use_dialect(mock_bigquery_client):
     pandas_gbq.read_gbq("SELECT 1")
 
     _, kwargs = mock_bigquery_client.query.call_args
+    assert not kwargs["job_config"].use_legacy_sql
+    pandas_gbq.context.dialect = None  # Reset the global state.
+
+
+def test_read_gbq_should_use_dialect_with_query_and_wait(
+    monkeypatch, mock_bigquery_client
+):
+    if not hasattr(mock_bigquery_client, "query_and_wait"):
+        pytest.skip(
+            f"google-cloud-bigquery {google.cloud.bigquery.__version__} does not have query_and_wait"
+        )
+
+    import pandas_gbq
+    import pandas_gbq.features
+
+    monkeypatch.setattr(
+        pandas_gbq.features.FEATURES,
+        "_bigquery_installed_version",
+        packaging.version.parse(pandas_gbq.features.BIGQUERY_QUERY_AND_WAIT_VERSION),
+    )
+
+    assert pandas_gbq.context.dialect is None
+    pandas_gbq.context.dialect = "legacy"
+    pandas_gbq.read_gbq("SELECT 1")
+
+    _, kwargs = mock_bigquery_client.query_and_wait.call_args
+    assert kwargs["job_config"].use_legacy_sql
+
+    pandas_gbq.context.dialect = "standard"
+    pandas_gbq.read_gbq("SELECT 1")
+
+    _, kwargs = mock_bigquery_client.query_and_wait.call_args
     assert not kwargs["job_config"].use_legacy_sql
     pandas_gbq.context.dialect = None  # Reset the global state.