@tswast 2024-01-10
I tested this manually in a local ipython instance to verify that the latency is improved when using this new functionality:

**After** this change:

```python
# In [1]:
import pandas_gbq

# In [2]: %%timeit -n10 -r10
df = pandas_gbq.read_gbq("SELECT 1", progress_bar_type=None)

# Out [2]:
# 394 ms ± 37.8 ms per loop (mean ± std. dev. of 10 runs, 10 loops each)
```

**Before** this change:

```python
# In [1]:
import pandas_gbq

# In [2]: %%timeit -n10 -r10
df = pandas_gbq.read_gbq("SELECT 1", progress_bar_type=None)

# Out [2]:
# 1.08 s ± 102 ms per loop (mean ± std. dev. of 10 runs, 10 loops each)
```

This is quite similar to the results I observed when implementing this optimization in the DB-API connector. https://github.com/googleapis/python-bigquery/pull/1747#issuecomment-1852368885