@tswast 2024-01-25
@Linchin Tests are passing after the most recent google-cloud-bigquery release. OK for another look.